move SupportedTechnologies to own package
