+++ de.atb.typhondl.xtext.ui/src/de/atb/typhondl/xtext/ui/creationWizard/CreationAnalyticsPage.java
@@ -35,9 +35,9 @@ import org.eclipse.swt.widgets.Text;
 
 import de.atb.typhondl.xtext.ui.modelUtils.ContainerService;
 import de.atb.typhondl.xtext.ui.properties.PropertiesService;
+import de.atb.typhondl.xtext.ui.technologies.SupportedTechnologies;
 import de.atb.typhondl.xtext.ui.utilities.EvolutionConfigEditor;
 import de.atb.typhondl.xtext.ui.utilities.InputField;
-import de.atb.typhondl.xtext.ui.utilities.SupportedTechnologies;
 
 /**
  * Optional page for the TyphonDL {@link CreateModelWizard}. The properties for
