+++ de.atb.typhondl.xtext.ui/src/de/atb/typhondl/xtext/ui/creationWizard/wizardPageAreas/VolumesDialog.java
@@ -41,7 +41,7 @@ import org.eclipse.xtext.conversion.ValueConverterException;
 import de.atb.typhondl.xtext.typhonDL.Volume_Properties;
 import de.atb.typhondl.xtext.ui.activator.Activator;
 import de.atb.typhondl.xtext.ui.modelUtils.VolumesService;
-import de.atb.typhondl.xtext.ui.utilities.SupportedTechnologies;
+import de.atb.typhondl.xtext.ui.technologies.SupportedTechnologies;
 
 /**
  * Dialog to add and edit Volumes in {@link VolumesArea}
