+++ de.atb.typhondl.xtext.ui/src/de/atb/typhondl/xtext/ui/modelUtils/ContainerService.java
@@ -39,7 +39,7 @@ import de.atb.typhondl.xtext.typhonDL.Services;
 import de.atb.typhondl.xtext.typhonDL.TyphonDLFactory;
 import de.atb.typhondl.xtext.typhonDL.URI;
 import de.atb.typhondl.xtext.typhonDL.Volumes;
-import de.atb.typhondl.xtext.ui.utilities.SupportedTechnologies;
+import de.atb.typhondl.xtext.ui.technologies.SupportedTechnologies;
 
 /**
  * Utility class for easier {@link Container} model object handling
