+++ de.atb.typhondl.xtext.ui/src/de/atb/typhondl/xtext/ui/modelUtils/ReplicationService.java
@@ -5,7 +5,7 @@ import java.util.Properties;
 import de.atb.typhondl.xtext.typhonDL.Modes;
 import de.atb.typhondl.xtext.typhonDL.Replication;
 import de.atb.typhondl.xtext.typhonDL.TyphonDLFactory;
-import de.atb.typhondl.xtext.ui.utilities.SupportedTechnologies;
+import de.atb.typhondl.xtext.ui.technologies.SupportedTechnologies;
 
 public class ReplicationService {
