+++ de.atb.typhondl.xtext.ui/src/de/atb/typhondl/xtext/ui/refactoring/ClusterTypeDialog.java
@@ -46,7 +46,7 @@ import de.atb.typhondl.xtext.typhonDL.Ports;
 import de.atb.typhondl.xtext.ui.modelUtils.ContainerService;
 import de.atb.typhondl.xtext.ui.modelUtils.ModelService;
 import de.atb.typhondl.xtext.ui.properties.PropertiesService;
-import de.atb.typhondl.xtext.ui.utilities.SupportedTechnologies;
+import de.atb.typhondl.xtext.ui.technologies.SupportedTechnologies;
 
 public class ClusterTypeDialog extends StatusDialog {
