+++ de.atb.typhondl.xtext.ui/src/de/atb/typhondl/xtext/ui/scriptGeneration/DeploymentModelService.java
@@ -55,7 +55,7 @@ import de.atb.typhondl.xtext.ui.modelUtils.ModelService;
 import de.atb.typhondl.xtext.ui.modelUtils.SoftwareService;
 import de.atb.typhondl.xtext.ui.modelUtils.VolumesService;
 import de.atb.typhondl.xtext.ui.properties.PropertiesService;
-import de.atb.typhondl.xtext.ui.utilities.SupportedTechnologies;
+import de.atb.typhondl.xtext.ui.technologies.SupportedTechnologies;
 
 public class DeploymentModelService {
