+++ de.atb.typhondl.xtext.ui/src/de/atb/typhondl/xtext/ui/scriptGeneration/GenerationService.java
@@ -41,8 +41,8 @@ import de.atb.typhondl.xtext.typhonDL.DeploymentModel;
 import de.atb.typhondl.xtext.typhonDL.Import;
 import de.atb.typhondl.xtext.ui.modelUtils.ModelService;
 import de.atb.typhondl.xtext.ui.properties.PropertiesService;
+import de.atb.typhondl.xtext.ui.technologies.SupportedTechnologies;
 import de.atb.typhondl.xtext.ui.utilities.PropertiesLoader;
-import de.atb.typhondl.xtext.ui.utilities.SupportedTechnologies;
 
 public class GenerationService {
